Migration to Visual Studio 2022

+++ SBFspot/SBFspot.cpp
@@ -2521,7 +2521,7 @@ E_SBFSPOT getInverterData(InverterData *device, unsigned long command, unsigned 
                                 {
                                     mppt new_mppt;
                                     new_mppt.Pdc(value);
-                                    device->mpp.insert(std::make_pair(cls, new_mppt));
+                                    device->mpp.insert(std::make_pair((uint8_t)cls, new_mppt));
                                 }
 
                                 debug_watt((std::string("SPOT_PDC") + std::to_string(cls)).c_str(), value, datetime);
@@ -2540,7 +2540,7 @@ E_SBFSPOT getInverterData(InverterData *device, unsigned long command, unsigned 
                                 {
                                     mppt new_mppt;
                                     new_mppt.Udc(value);
-                                    device->mpp.insert(std::make_pair(cls, new_mppt));
+                                    device->mpp.insert(std::make_pair((uint8_t)cls, new_mppt));
                                 }
 
                                 debug_volt((std::string("SPOT_UDC") + std::to_string(cls)).c_str(), value, datetime);
@@ -2557,7 +2557,7 @@ E_SBFSPOT getInverterData(InverterData *device, unsigned long command, unsigned 
                                 {
                                     mppt new_mppt;
                                     new_mppt.Idc(value);
-                                    device->mpp.insert(std::make_pair(cls, new_mppt));
+                                    device->mpp.insert(std::make_pair((uint8_t)cls, new_mppt));
                                 }
 
                                 debug_amp((std::string("SPOT_IDC") + std::to_string(cls)).c_str(), value, datetime);
@@ -2994,8 +2994,8 @@ void resetInverterData(InverterData *inv)
     inv->MeteringGridMsTotWIn = 0;
     inv->MeteringGridMsTotWOut = 0;
     inv->hasBattery = false;
-    inv->mpp.insert(std::make_pair(1, mppt(0, 0, 0)));
-    inv->mpp.insert(std::make_pair(2, mppt(0, 0, 0)));
+    inv->mpp.insert(std::make_pair((uint8_t)1, mppt(0, 0, 0)));
+    inv->mpp.insert(std::make_pair((uint8_t)2, mppt(0, 0, 0)));
 }
 
 E_SBFSPOT setDeviceData(InverterData *inv, LriDef lri, uint16_t cmd, Rec40S32 &data)
